trackpoint-backend/requirements.txt
fastapi==0.115.5
fastapi_boot==0.0.28
pydantic==2.9.1
PyJWT==2.10.1
PyYAML==6.0.2
redis==5.2.0
starlette==0.41.3
itsdangerous==2.0.1
tortoise-orm==0.21.6  # 坑，pipreqs生成的是tortoise，是另一个无关的库
pydantic[email]
typing_extensions==4.12.2
uvicorn==0.32.1
aiomysql==0.2.0
minio==7.2.12
python-multipart==0.0.20
cryptography==42.0.2